refactor : IndexInfo 엔티티 baseIndex를 BigDecimal로 변경

+++ src/main/java/com/codeit/team7/findex/domain/entity/IndexInfo.java
@@ -7,6 +7,7 @@ import jakarta.persistence.EnumType;
 import jakarta.persistence.Enumerated;
 import jakarta.persistence.Table;
 import jakarta.persistence.UniqueConstraint;
+import java.math.BigDecimal;
 import java.time.LocalDate;
 import lombok.AccessLevel;
 import lombok.Getter;
@@ -25,26 +26,34 @@ public class IndexInfo extends BaseUpdatableEntity {
 
   @Column(nullable = false, name = "idx_csf", length = 300)
   private String indexClassification;
+
   @Column(nullable = false, name = "idx_nm", length = 200)
   private String indexName;
+
   @Column(nullable = false, name = "itms_cnt")
   private Integer itemCount;
+
   @Column(nullable = false, name = "bas_pntm")
   private LocalDate basePointInTime;
+
+  // 소수점 사용 -> db 타입 변경 필요
   @Column(nullable = false, name = "bas_idx")
-  private Integer baseIndex;
+  private BigDecimal baseIndex;
+
   @Enumerated(EnumType.STRING)
   @Column(nullable = false, name = "source_type", length = 20)
   private SourceType sourceType;
+
   // 즐겨찾기 기본값 false
   @Column(nullable = false, name = "favorite")
   private boolean favorite;
+
   // 활성화 기본값 true
   @Column(nullable = false, name = "enabled")
   private boolean enabled = true;
 
   public IndexInfo(String indexClassification, String indexName, Integer itemCount,
-      LocalDate basePointInTime, Integer baseIndex, SourceType sourceType, boolean favorite,
+      LocalDate basePointInTime, BigDecimal baseIndex, SourceType sourceType, boolean favorite,
       boolean enabled) {
     this.indexClassification = indexClassification;
     this.indexName = indexName;
@@ -56,7 +65,7 @@ public class IndexInfo extends BaseUpdatableEntity {
     this.enabled = enabled;
   }
 
-  public void update(Integer newItemCount, LocalDate newBasePointInTime, Integer newBaseIndex,
+  public void update(Integer newItemCount, LocalDate newBasePointInTime, BigDecimal newBaseIndex,
       boolean newFavorite) {
     if (newItemCount != null && !newItemCount.equals(this.itemCount)) {
       this.itemCount = newItemCount;
